+++ lib/operations/add_user.js
@@ -7,18 +7,11 @@ const CONSTANTS = require('../constants');
 const count = require('./collection_ops').count;
 const crypto = require('crypto');
 const handleCallback = require('../utils').handleCallback;
+const loadDb = require('../dynamic_loaders').loadDb;
 const MongoError = require('mongodb-core').MongoError;
 const ReadPreference = require('mongodb-core').ReadPreference;
 const toError = require('../utils').toError;
 const updateOne = require('./collection_ops').updateOne;
-
-let db;
-function loadDb() {
-  if (!db) {
-    db = require('../db');
-  }
-  return db;
-}
 
 class AddUserOperation extends OperationBase {
   constructor(db, username, password, options) {
