+++ lib/operations/common_functions.js
@@ -7,17 +7,10 @@ const decorateWithReadConcern = require('../utils').decorateWithReadConcern;
 const executeCommand = require('./db_ops').executeCommand;
 const formattedOrderClause = require('../utils').formattedOrderClause;
 const handleCallback = require('../utils').handleCallback;
+const loadCursor = require('../dynamic_loaders').loadCursor;
 const MongoError = require('mongodb-core').MongoError;
 const ReadPreference = require('mongodb-core').ReadPreference;
 const toError = require('../utils').toError;
-
-let cursor;
-function loadCursor() {
-  if (!cursor) {
-    cursor = require('../cursor');
-  }
-  return cursor;
-}
 
 /**
  * Build the count command.
@@ -225,11 +218,12 @@ function prepareDocs(coll, docs, options) {
 function nextObject(cursor, callback) {
   let Cursor = loadCursor();
 
-  if (cursor.s.state === Cursor.CLOSED || (cursor.isDead && cursor.isDead()))
+  if (cursor.s.state === Cursor.CLOSED || (cursor.isDead && cursor.isDead())) {
     return handleCallback(
       callback,
       MongoError.create({ message: 'Cursor is closed', driver: true })
     );
+  }
   if (cursor.s.state === Cursor.INIT && cursor.s.cmd.sort) {
     try {
       cursor.s.cmd.sort = formattedOrderClause(cursor.s.cmd.sort);
