+++ lib/operations/create_collection.js
@@ -4,16 +4,9 @@ const OperationBase = require('./operation').OperationBase;
 const CommandOperation = require('./command');
 const applyWriteConcern = require('../utils').applyWriteConcern;
 const handleCallback = require('../utils').handleCallback;
+const loadCollection = require('../dynamic_loaders').loadCollection;
 const MongoError = require('mongodb-core').MongoError;
 const ReadPreference = require('mongodb-core').ReadPreference;
-
-let collection;
-function loadCollection() {
-  if (!collection) {
-    collection = require('../collection');
-  }
-  return collection;
-}
 
 // Filter out any write concern options
 const illegalCommandFields = [
