+++ lib/operations/has_next.js
@@ -2,16 +2,9 @@
 
 const Aspect = require('./operation').Aspect;
 const defineAspects = require('./operation').defineAspects;
+const loadCursor = require('../dynamic_loaders').loadCursor;
 const OperationBase = require('./operation').OperationBase;
 const nextObject = require('./common_functions').nextObject;
-
-let cursor;
-function loadCursor() {
-  if (!cursor) {
-    cursor = require('../cursor');
-  }
-  return cursor;
-}
 
 class HasNextOperation extends OperationBase {
   constructor(cursor) {
