+++ lib/operations/map_reduce.js
@@ -7,17 +7,10 @@ const decorateWithReadConcern = require('../utils').decorateWithReadConcern;
 const executeCommand = require('./db_ops').executeCommand;
 const handleCallback = require('../utils').handleCallback;
 const isObject = require('../utils').isObject;
+const loadDb = require('../dynamic_loaders').loadDb;
 const OperationBase = require('./operation').OperationBase;
 const resolveReadPreference = require('../utils').resolveReadPreference;
 const toError = require('../utils').toError;
-
-let db;
-function loadDb() {
-  if (!db) {
-    db = require('../db');
-  }
-  return db;
-}
 
 const exclusionList = ['readPreference', 'session', 'bypassDocumentValidation'];
