+++ lib/operations/operation.js
@@ -41,7 +41,7 @@ class OperationBase {
 }
 
 function defineAspects(operation, aspects) {
-  if (!Array.isArray(aspects)) {
+  if (!Array.isArray(aspects) && !(aspects instanceof Set)) {
     aspects = [aspects];
   }
   aspects = new Set(aspects);
