+++ lib/operations/remove_user.js
@@ -6,17 +6,10 @@ const applyWriteConcern = require('../utils').applyWriteConcern;
 const CONSTANTS = require('../constants');
 const findOne = require('./collection_ops').findOne;
 const handleCallback = require('../utils').handleCallback;
+const loadDb = require('../dynamic_loaders').loadDb;
 const MongoError = require('mongodb-core').MongoError;
 const ReadPreference = require('mongodb-core').ReadPreference;
 const remove = require('./collection_ops').remove;
-
-let db;
-function loadDb() {
-  if (!db) {
-    db = require('../db');
-  }
-  return db;
-}
 
 class RemoveUserOperation extends OperationBase {
   constructor(db, username, options) {
