+++ lib/operations/rename.js
@@ -5,15 +5,8 @@ const applyWriteConcern = require('../utils').applyWriteConcern;
 const checkCollectionName = require('../utils').checkCollectionName;
 const executeDbAdminCommand = require('./db_ops').executeDbAdminCommand;
 const handleCallback = require('../utils').handleCallback;
+const loadCollection = require('../dynamic_loaders').loadCollection;
 const toError = require('../utils').toError;
-
-let collection;
-function loadCollection() {
-  if (!collection) {
-    collection = require('../collection');
-  }
-  return collection;
-}
 
 class RenameOperation extends OperationBase {
   constructor(collection, newName, options) {
