+++ lib/operations/to_array.js
@@ -3,16 +3,9 @@
 const Aspect = require('./operation').Aspect;
 const defineAspects = require('./operation').defineAspects;
 const handleCallback = require('../utils').handleCallback;
+const loadCursor = require('../dynamic_loaders').loadCursor;
 const OperationBase = require('./operation').OperationBase;
 const push = Array.prototype.push;
-
-let cursor;
-function loadCursor() {
-  if (!cursor) {
-    cursor = require('../cursor');
-  }
-  return cursor;
-}
 
 class ToArrayOperation extends OperationBase {
   constructor(cursor) {
